requirements.txt
Django==1.11
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.7
yapf==0.24.0
